+++ .gitignore
@@ -1,3 +1,4 @@
+packages/container
 *.bpkg
 *.gem
 *.rbc
