+++ package.json
@@ -1,6 +1,8 @@
 {
   "devDependencies": {
     "defeatureify": "~0.2.0",
-    "yuidocjs": "~0.3.46"
+    "yuidocjs": "~0.3.46",
+    "es6-module-transpiler": "~0.3.6",
+    "glob": "~3.2.8"
   }
 }
