+++ packages/ember-runtime/lib/system/container.js
@@ -1,2 +1,2 @@
-Ember.Container = requireModule('container');
+Ember.Container = requireModule('container')['default'];
 Ember.Container.set = Ember.set;
